+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - possibility to filter tasks from queue
 - exported EVENTS object with list of supported events
 - cancelling group of tasks by predicate
+- ability to get task position in the queue
+- ability to check if task is currently running
 ### Changed
 - transpiling options, now transpiled code is for environments used by at least 3% of people
 ### Fixed